Config handling changes for #191.
Change adds a config.CONFIG_MSG text blob that reminds the user to copy
the source config out and edit it, but we're keeping the source fallback
config for now (due to #446; it's needed for tests to run).

turbiniactl changes to print the friendlier message instead of the
traceback on failure to load config, but we include the traceback
exception message for visibility.